requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
praw==7.7.1
prawcore==2.4.0
vaderSentiment==3.3.2
python-dotenv==1.0.1
pydantic==2.9.2
